+++ api/src/main/java/io/strimzi/api/kafka/model/template/PodTemplate.java
@@ -52,6 +52,7 @@ public class PodTemplate implements Serializable, UnknownPropertyPreserving {
     private String priorityClassName;
     private String schedulerName;
     private List<HostAlias> hostAliases;
+    private Boolean enableServiceLinks;
     private Map<String, Object> additionalProperties = new HashMap<>(0);
 
     @Description("Metadata applied to the resource.")
@@ -160,7 +161,7 @@ public class PodTemplate implements Serializable, UnknownPropertyPreserving {
     }
 
     @Description("The pod's HostAliases. " +
-            "HostAliases is an optional list of hosts and IPs that will be injected into the pod's hosts file if specified.")
+            "HostAliases is an optional list of hosts and IPs that will be injected into the Pod's hosts file if specified.")
     @KubeLink(group = "core", version = "v1", kind = "HostAlias")
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     public List<HostAlias> getHostAliases() {
@@ -169,6 +170,16 @@ public class PodTemplate implements Serializable, UnknownPropertyPreserving {
 
     public void setHostAliases(List<HostAlias> hostAliases) {
         this.hostAliases = hostAliases;
+    }
+
+    @Description("Indicates whether information about services should be injected into Pod's environment variables.")
+    @JsonInclude(JsonInclude.Include.NON_DEFAULT)
+    public Boolean getEnableServiceLinks() {
+        return enableServiceLinks;
+    }
+
+    public void setEnableServiceLinks(Boolean enableServiceLinks) {
+        this.enableServiceLinks = enableServiceLinks;
     }
 
     @Override